set modules
