+++ src/modules/add.js
@@ -1,0 +1,26 @@
+const error = document.getElementById('error');
+const addName = async (user, score) => {
+  if (user === '' || score === '') {
+    error.innerHTML = 'Please fill all the required fields';
+    setTimeout(() => { error.innerHTML = ''; }, 3000);
+  } else {
+    const response = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/4tfml0QLmgKTloHBYIf5/scores/', {
+      method: 'POST',
+      body: JSON.stringify({
+        user,
+        score,
+      }),
+      headers: {
+        'Content-type': 'application/json; charset=UTF-8',
+      },
+    });
+    if (response.status !== 201) {
+      error.innerHTML = 'Something went wrong';
+      setTimeout(() => { error.innerHTML = ''; }, 3000);
+    }
+    document.getElementById('inputName').value = '';
+    document.getElementById('inputScore').value = '';
+  }
+};
+
+export default addName;
